Move BackgroundRequestLooper to its own file.
